feat: don't use `Option` as poor man's optional parameter
It's a complete waste of computing resources. It doesn't contribute to
readability anyway. (Take `x.set_html(something, None)` as an example.
The readers can't tell which portion of `x` this statement intends to
update. On the other hand, `x.set_html(something, SIH_REPLACE_CONTENT)`
is verbose but sufficiently clear.)